security system using TM4C123GH6PM
This project is based on home automation .In today's world security has been  major issue ,so this project will  give as quick prototype on building the security system.
 
This project involves TM4C123GH6PM microcontroller board as it's core, which is manufactured by texas instruments and consists software as tivaware .

This project also involves  2X16 LCD display  screen to show the status of the accessing i,e correct or wrong password, here we require a 10k pot to adjust the contrast of LCD

We enter the password by QWERTY keyboard and its get loaded in memory of microcontroller. we get "*" symbol while typing password on LCD. here in code we set password as 4 digit number ,after entering password it is compared by default password if it is matched then we get "correct" message on LCD and green light is glown. else a "wrong" message is displayed and red light is glown.